- refactor, moved image view code into ImageViewer from Editormanager - remove obsolete clean up in ProjectManager (+3 squashed commits) -  remove unused var -  fix coloring of image in project
